# .kilo/worktrees/different-mushroom/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ИТМО — ИИ-ассистент для проверки соответствия критериям учёного звания
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /criteria:
    get:
      operationId: getCriteria
      tags: [criteria]
      summary: Получить актуальные критерии
      parameters:
        - in: query
          name: rankType
          schema:
            type: string
            enum: [docent, professor]
      responses:
        "200":
          description: Список критериев
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CriteriaRule"

    post:
      operationId: createCriteria
      tags: [criteria]
      summary: Создать критерии (только admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCriteriaRequest"
      responses:
        "201":
          description: Критерии созданы
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CriteriaRule"
  /criteria/{id}:
    put:
      operationId: updateCriteria
      tags: [criteria]
      summary: Обновить критерии (только admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCriteriaRequest"
      responses:
        "200":
          description: Критерии обновлены
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CriteriaRule"

  # === APPLICATIONS ===
...
  /chat/messages:
    get:
      operationId: getChatHistory
      tags: [chat]
      summary: История чата пользователя
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: История сообщений
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"

    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Отправить вопрос ИИ-ассистенту
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendChatMessageRequest"
      responses:
        "200":
          description: Ответ ИИ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
  /chat/messages/{id}/rate:
    post:
      operationId: rateChatMessage
      tags: [chat]
      summary: Оценить ответ ИИ
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RateMessageRequest"
      responses:
        "200":
          description: Оценка сохранена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"

  # === DOCUMENTS ===
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CriteriaRule:
      type: object
      properties:
        id:
          type: integer
        rankType:
          type: string
          enum: [docent, professor]
        version:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        rules:
          $ref: "#/components/schemas/CriteriaRules"
      required:
        - id
        - rankType
        - version
        - isActive
        - createdAt
        - rules
    CriteriaRules:
      type: object
      properties:
        academicExperienceYears:
          type: integer
          description: Стаж научно-педагогической работы (лет)
        pedagogicalExperienceYears:
          type: integer
          description: Стаж педагогической работы по специальности (лет)
        publicationsCount:
          type: integer
          description: Публикации в рецензируемых изданиях за 5 лет
        textbooksCount:
          type: integer
          description: Учебные издания
        scopusWosCount:
          type: integer
          description: Публикации в Scopus/WoS
        requiredDegree:
          type: string
          description: Требуемая учёная степень (candidate/doctor)
        graduatesCount:
          type: integer
          description: Подготовленные аспиранты (для профессора)
          nullable: true
      required:
        - academicExperienceYears
        - pedagogicalExperienceYears
        - publicationsCount
        - textbooksCount
        - scopusWosCount
        - requiredDegree
    CreateCriteriaRequest:
      type: object
      properties:
        rankType:
          type: string
          enum: [docent, professor]
        rules:
          $ref: "#/components/schemas/CriteriaRules"
      required:
        - rankType
        - rules

    # === APPLICATION SCHEMAS ===
...
    ChatMessage:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        message:
          type: string
        response:
          type: string
        contextAppId:
          type: integer
          nullable: true
        rating:
          type: integer
          nullable: true
          description: 1 = полезно, -1 = не полезно
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - message
        - response
        - createdAt
    SendChatMessageRequest:
      type: object
      properties:
        message:
          type: string
          minLength: 1
        contextAppId:
          type: integer
          nullable: true
      required:
        - message
    RateMessageRequest:
      type: object
      properties:
        rating:
          type: integer
          enum: [1, -1]
      required:
        - rating

    # === DOCUMENT SCHEMAS ===
...